:bug: fix(voucher-print-setup): Fix voucher print logo visibility issue
:pencil: Messages :
This commit addresses a bug where the logo in the voucher print setup was not being displayed. The modifications include changes in the VoucherBatchController, and the livewire components responsible for the setup of the voucher print and its logo.
Additionally, related database seeders have been updated.
The composer.lock file was also modified as part of this commit.